Use C++14 (#1836)
* 1835: use --with-cxx=14 when building wxwidgets

* 1835: raise CMAKE_CXX_STANDARD to 14

* 1835: use an auto return type for findSubtrahendPlanes()

* 1835: travis-linux.sh: use gcc5

* 1835: fix some implict conversion warnings with clang

* 1835: Linux: print results of "ldd" on the trenchbroom executable

* 1835: travis-linux.sh: use ldd --verbose to see library versions

+++ common/src/Renderer/ShaderProgram.cpp
@@ -138,7 +138,7 @@ namespace TrenchBroom {
                 GLint infoLogLength = 0;
                 glAssert(glGetProgramiv(m_programId, GL_INFO_LOG_LENGTH, &infoLogLength));
                 if (infoLogLength > 0) {
-                    char* infoLog = new char[infoLogLength];
+                    char* infoLog = new char[static_cast<size_t>(infoLogLength)];
                     glAssert(glGetProgramInfoLog(m_programId, infoLogLength, &infoLogLength, infoLog));
                     infoLog[infoLogLength-1] = 0;
